This is my code link in Google Colab.
(1) Build a PQC with data re-uploading
At the core of both RL algorithms we are implementing is a PQC that takes as input the
agent's state s in the environment (i.e., a numpy array) and outputs a vector of expectation
values. These expectation values are then post-processed, to produce an agent's policy or
approximate Q-values. In this way, the PQCs are playing an analog role to that of deep neural
networks in modern deep RL algorithms.
[8]
A popular way to encode an input vector in a PQC is through the use of single-qubit rotations, where rotation angles are controlled by the components of this input vector. In order to get
a highly-expressive model, these single-qubit encodings are not performed only once in the
PQC, but in several "re-uploadings", interlayed with variational gates.
[9]
To further enhance the expressivity and trainability of data, re-uploading PQCs is to use
trainable input-scaling parameters for each encoding gate of the PQC, and trainable
observable weights w at its output.
[10] Then, I will start by constructing, out of the PQC that
was just defined, the softmax-VQC policy (where VQC stands for variational quantum
circuit). To implement the weighting of the Pauli product, I can use an
extra tf.keras.layers.Layer that stores the action-specific weights and applies them
multiplicatively on the expectation value. (2) Decide the best strategy of agent X(Teacher)
For the game of tic-tac-toe, because its board is a simple 3x3 grid, people already know the
optimal strategy for the game. That is to say, for each possible game state, we know the best
move to make as human. Therefore, We set the agent X as a teacher who knows the optimal
strategy, or the best player. However, we should also make sure that the teacher takes
10% random actions, in addition to making optimal moves, because if the teacher acts as the
best strategy every time, the QRL agent will not be able to win and learn effectively. Then the
agent X can be a good teacher of Quantum Reinforcement Learning agent. QRL agent can
learn to win its teacher and at the same time it can learn to be better!
(3) Create the Tic-Tac-Toe env similar to gym env
To be suitable to the existed code about the quantum circuit and deep-learning model, we
need to create the tic-tac-toe env by ourselves. The env need to be similar to the gym env, 4
Quantum Reinforcement Learning Model
for Tic-Tac-Toe Environment
which has its variables, such as, step, action, reward and so on. For the tic-tac-toe env, we set
two agents including O and X. X is a teacher, who acts as 90% best strategy and O is
Quantum Reinforcement Learning agent, whose strategy is trained by Quantum
Reinforcement Learning Model. After that, we can use our quantum reinforcement learning
model for Tic-Tac-Toe env and train the better strategy in the env according to the reward.
[11]
(4) Train the quantum reinforcement learning model
Define a function that gathers episodes of interaction with the environment and a function that
computes discounted returns out of the rewards collected in an episode.
[12] Define the
hyperparameters and prepare the optimizers. Implement a function that updates the policy
using states, actions and returns. In the end, we can implement the main training loop of the
agent. The experience code is save in colab, and the link is as follow:
https://colab.research.google.com/drive/15rb0vnUIGK1TPK2Kh8u5k9JWlv14uLWM#scrollT
o=bfqZ5HLHyqU3